% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descrip.R
\docType{data}
\name{coi1}
\alias{coi1}
\title{Table of 1-to-1 coicop groups
Data containing coicop groups where there is a 1-to-1 conversion between ecoicop and the 2018 UN coicop.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 217 rows and 2 columns.
}
\usage{
data(coi1)
}
\description{
Table of 1-to-1 coicop groups
Data containing coicop groups where there is a 1-to-1 conversion between ecoicop and the 2018 UN coicop.
}
\keyword{datasets}
