% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descrip.R
\docType{data}
\name{coi_descrp}
\alias{coi_descrp}
\title{UN coicop description
New 2018 coicop groups with Norwegian and English names (Unofficial).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 298 rows and 4 columns.
}
\usage{
data(coi_descrp)
}
\description{
UN coicop description
New 2018 coicop groups with Norwegian and English names (Unofficial).
}
\keyword{datasets}
