% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coicop_convert_funcs.R
\name{coicop_convert}
\alias{coicop_convert}
\title{Convert from European COICOP to UN COICOP}
\usage{
coicop_convert(
  data,
  ecoicop,
  varenavn = NULL,
  coicop6 = NULL,
  ingredients = NULL,
  maxtemp = NULL
)
}
\arguments{
\item{data}{Name of the data set}

\item{ecoicop}{Name of the variable for the European COICOP}

\item{varenavn}{Name of the variable for the product name(s)}

\item{coicop6}{Name of the variable containing COICOP6}

\item{ingredients}{Name of the variable for the ingredients}

\item{maxtemp}{Name on the variable for the maximum storage temperature.}
}
\value{
UN COICOP (2018) classification
}
\description{
Convert from European COICOP to UN COICOP
}
