% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coicop_convert_funcs.R
\name{coicop_convert_single}
\alias{coicop_convert_single}
\title{Convert a single European COICOP to 2018 UN COICOP}
\usage{
coicop_convert_single(
  ecoicop,
  coi1,
  coicop_ref,
  varenavn = NULL,
  coicop6 = NULL,
  ingredients = NULL,
  maxtemp = NULL
)
}
\arguments{
\item{ecoicop}{String contain the COICOP (5-digits, no periods)}

\item{coi1}{The 1:1 table}

\item{coicop_ref}{The reference data}

\item{varenavn}{String with the product name}

\item{coicop6}{String with the COICOP 6 classification (if available)}

\item{ingredients}{String with the list of ingredients (if available)}

\item{maxtemp}{Maximum temperature for storage (if available)}
}
\value{
string with COICOP value
}
\description{
Convert a single European COICOP to 2018 UN COICOP
}
\keyword{internal}
