% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descrip.R
\docType{data}
\name{coicop_ref}
\alias{coicop_ref}
\title{Reference table
Table containing key words and conditions for determining new coicop groups.}
\format{
An object of class \code{data.frame} with 52 rows and 8 columns.
}
\usage{
data(coicop_ref)
}
\description{
Reference table
Table containing key words and conditions for determining new coicop groups.
}
\keyword{datasets}
