% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coicop_convert_funcs.R
\name{intern_convert}
\alias{intern_convert}
\title{Lower level internal convert function}
\usage{
intern_convert(
  coicop_words,
  ecoicop,
  coicop18,
  coicop6 = NULL,
  coicop6_test = NULL,
  makstemp = NULL,
  makstemp_limit = NULL,
  varenavn = NULL,
  ingredients = NULL,
  ingredients_test = FALSE
)
}
\arguments{
\item{coicop_words}{String with words to identify in the product name}

\item{ecoicop}{String contain the COICOP (5-digits, no periods)}

\item{coicop18}{Vector of two: first is the COICOP to allocated given words,coicop 6,max temp found; second is the alternate.}

\item{coicop6}{String with the COICOP 6 classification (if available)}

\item{coicop6_test}{String with Coicop 6 group or groups to check for.}

\item{makstemp}{Maximum temperature for storage (if available)}

\item{makstemp_limit}{Maximum temperature to check for.}

\item{varenavn}{String with the product name}

\item{ingredients}{string with list of ingredients}

\item{ingredients_test}{Whether to check ingredients}
}
\value{
String with new coicop group
}
\description{
Lower level internal convert function
}
