% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descrip.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test data for food products
Example data for using in testing and examples}
\format{
An object of class \code{matrix} with 5 rows and 6 columns.
}
\usage{
data(testdata)
}
\description{
Test data for food products
Example data for using in testing and examples
}
\keyword{datasets}
